Refactoring the vertex selection algorithms.
